+++ src/main/java/com/korit/carecheckkoreait/controller/AuthController.java
@@ -1,0 +1,33 @@
+package com.korit.carecheckkoreait.controller;
+
+import com.korit.carecheckkoreait.dto.request.ReqSigninDto;
+import com.korit.carecheckkoreait.dto.response.RespTokenDto;
+import com.korit.carecheckkoreait.service.UserService;
+import io.swagger.v3.oas.annotations.Operation;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequestMapping("/auth")
+public class AuthController {
+
+    @Autowired
+    private UserService userService;
+
+    @Operation(summary = "로그인", description = "로그인")
+    @PostMapping("/signin")
+    public ResponseEntity<?> signin(@RequestBody ReqSigninDto dto) {
+        RespTokenDto respTokenDto = RespTokenDto.builder()
+                .type("JWT")
+                .name("AccessToken")
+                .token(userService.signin(dto))
+                .build();
+        System.out.println(respTokenDto);     // 테스트
+
+        return ResponseEntity.ok().body(respTokenDto);
+    }
+}
